chore: move well-known test addresses outside deployed code.
