+++ src/logion/services/authentication.service.ts
@@ -2,7 +2,6 @@ import jwt, { Algorithm, Jwt, VerifyErrors } from "jsonwebtoken";
 import { injectable } from "inversify";
 import { Request } from "express";
 import { UnauthorizedException } from "dinoloop/modules/builtin/exceptions/exceptions";
-import { ALICE, BOB, CHARLY } from "../model/addresses.model";
 import moment, { Moment } from "moment";
 
 const ALGORITHM: Algorithm = "HS384";
@@ -22,7 +21,7 @@ export class LogionUserCheck implements AuthenticatedUser {
         this.nodeOwner = nodeOwner;
     }
 
-    private nodeOwner: string;
+    private readonly nodeOwner: string;
 
     requireLegalOfficer(): void {
         this.require(user => user.legalOfficer, "Reserved to legal officer");
@@ -161,7 +160,7 @@ export class AuthenticationService {
     }
 
     private isLegalOfficer(address: string): boolean {
-        return address === ALICE || address === BOB || address === CHARLY;
+        return address === this.nodeOwner;
     }
 
     private _unauthorized(error: VerifyErrors): UnauthorizedException<{ error: string }> {
