+++ test/unit/model/locrequest.model.spec.ts
@@ -1,5 +1,5 @@
 import { v4 as uuid } from "uuid";
-import { ALICE } from "../../../src/logion/model/addresses.model";
+import { ALICE } from "../../helpers/addresses";
 import moment, { Moment } from "moment";
 import {
     LocRequestDescription,
